Update Building a Medallion Architecture
In the following dataset we perform the following using pyspark  in the Microsoft Fabric platform. We Building two layers of data, a bronze and gold tables. We start of by extracting a file name 'Sales' consisting of one big table  into a Fabric Lakehouse. Read the file into a dataframe. We create dimension tables and one fact table . Each Table is created on its own notebook

Data is synced from the bronze layer tables to the gold layer tables . We finally create a mount notebook to mount into the filesystem.
All the notebooks are then called from a dedicated notebook. That notebook is put inside a pipeline to run on given times and conditions. We also produce a semantic model of the data. By completion, we have created a complete ETL process from datasource, doing transformations and finally loading into gold layer. 

We finally create a Power BI Visual for the data.