Update EarthQuake_end_to_end
In the following dataset is extracted from an API and dumped into a dataframe. We clean the dataset, create  fact & dimension tables, create timestamp columns. This is done through three notebooks. A bronze notebook, silver notebook, gold notebook is created for a pipeline. The pipeline is set to run the three notebooks, the bronze notebook, silver notebook, gold notebook.
 A semantic model by default is automatically created for us by fabric. We connect the semantic model to power bi through the sql analytics end point. By doing this, we keep the data in power bi report sync with the one in the lakehouse.
We finally build a power bi report to visualize this data. Everytime we refresh the pipeline, recent data is load through the pipeline into the lakehouse and visualized on power bi.